Compute SIMD direction-buffer size in 64-bit
dirbuffersize = s->qlen * s->maxdlen * 4 was evaluated in 32-bit int
(both qlen and maxdlen are int) before being stored in the uint64_t
result, so the multiplication could overflow (signed UB) once
qlen * maxdlen * 4 exceeds INT_MAX, i.e. a length product above
~5.37e8. This is well outside the current limits (the product is capped
at 25e6 and the length sum at 65535), so it is latent today, but the
narrowing would silently bite if the product limit were ever raised.

Cast the first operand to uint64_t so the whole expression is evaluated
in 64-bit. Behaviour is unchanged for all inputs the limits currently
admit (the value is identical whenever the int form does not overflow).

+++ src/align_simd.cc
@@ -948,7 +948,7 @@ auto backtrack16(s16info_s * s,
                  unsigned short * pgaps) -> void
 {
   unsigned short * dirbuffer = s->dir;
-  uint64_t const dirbuffersize = s->qlen * s->maxdlen * 4;
+  uint64_t const dirbuffersize = static_cast<uint64_t>(s->qlen) * s->maxdlen * 4;
   uint64_t const qlen = s->qlen;
   char const * qseq = s->qseq;
 
@@ -1391,7 +1391,7 @@ auto search16(s16info_s * s,
     }
   maxdlen = 4 * ((maxdlen + 3) / 4);
   s->maxdlen = maxdlen;
-  uint64_t const dirbuffersize = s->qlen * s->maxdlen * 4;
+  uint64_t const dirbuffersize = static_cast<uint64_t>(s->qlen) * s->maxdlen * 4;
 
   if (dirbuffersize > s->diralloc)
     {
